Add tasks to start containers for testing
Users shouldn't have to deal with *any* complexity for running tests.
Thus, tasks that perform the necessary commands are provided.
No knowledge of docker is now required, just an install and
knowledge that the tasks exist.

Further added env variables which tests will use to connect to the
various containers.